gaming

+++ DankleC/libc/include/dankle.h
@@ -5,6 +5,6 @@
 #define WRITE_CHAR(c) *((char*)0xFFFFFFF0) = c
 
 void println(const char* txt);
-char* itoa(int num, char* str, int base);
+char* itoa(int num, char* str, short base);
 
 #endif
